CTradeBase initial commit
